recorder/requirements.txt
numpy
opencv-python-headless
grpcio
icvproto
python-keycloak
opencv-python==4.2.0.32
ultralytics